// client/src/types/user.ts
export interface User {
    _id: string;
    username: string;
    email?: string;
    avatar: string;
    bio?: string;
    xp: number;
    level: number;
    levelBadge: string;
    createdAt?: string;
}

export interface RegisterPayload {
    username: string;
    email: string;
    password: string;
}

export interface LoginPayload {
    email: string;
    password: string;
}